progress on init refactor
