+++ example/lib/application_view_model.dart
@@ -1,18 +1,17 @@
-import 'package:empire/empire_property.dart';
+import 'package:empire/empire_properties.dart';
 import 'package:empire/empire_view_model.dart';
 import 'package:flutter/material.dart';
 import 'dart:math' as math;
 
 class ApplicationViewModel extends EmpireViewModel {
-  late final EmpireProperty<Color> backgroundColor;
-  late final EmpireStringProperty title;
+  final backgroundColor = EmpireProperty(Colors.white);
+
+  final EmpireStringProperty title;
+
+  ApplicationViewModel(this.title);
 
   @override
-  void initProperties() {
-    backgroundColor =
-        createProperty(Colors.white, propertyName: 'backgroundColor');
-    title = createStringProperty('Empire - Counter App Example');
-  }
+  Iterable<EmpireProperty> get props => [backgroundColor, title];
 
   void changeBackgroundColor(Color color) => backgroundColor(color);
