+++ lib/empire.dart
@@ -1,7 +1,3 @@
-///A simple yet powerful state management library for [Flutter](https://flutter.dev/).
 library empire;
 
-export 'empire_state.dart';
-export 'empire_view_model.dart';
-export 'empire_widget.dart';
-export 'empire_property.dart';
+export '/src/empire.dart';
