+++ lib/empire_properties.dart
@@ -1,0 +1,3 @@
+library empire;
+
+export 'src/empire_property.dart';
