+++ lib/src/empire_exceptions.dart
