+++ lib/src/empire_string_property.dart
@@ -5,7 +5,7 @@ part of 'empire_property.dart';
 ///When the value of this changes, it will send a [EmpireStateChanged] event by default. This includes
 ///automatically triggering a UI rebuild.
 class EmpireStringProperty extends EmpireProperty<String> {
-  EmpireStringProperty(super.value, super.viewModel, {super.propertyName});
+  EmpireStringProperty(super.value, {super.propertyName});
 
   ///Whether the string value is empty
   bool get isEmpty => _value.isEmpty;
@@ -57,8 +57,7 @@ class EmpireStringProperty extends EmpireProperty<String> {
 ///When the value of this changes, it will send a [EmpireStateChanged] event by default. This includes
 ///automatically triggering a UI rebuild.
 class EmpireNullableStringProperty extends EmpireProperty<String?> {
-  EmpireNullableStringProperty(super.value, super.viewModel,
-      {super.propertyName});
+  EmpireNullableStringProperty(super.value, {super.propertyName});
 
   ///Whether the string value is empty
   ///
@@ -109,4 +108,36 @@ class EmpireNullableStringProperty extends EmpireProperty<String?> {
   /// Both [start] and [end] must be non-negative and no greater than [length];
   /// [end], if provided, must be greater than or equal to [start].
   String? substring(int start, [int? end]) => _value?.substring(start, end);
+}
+
+///Short hand helper function for initializing an [EmpireStringProperty].
+///
+///See [EmpireProperty] for [propertyName] usages.
+///
+///## Example
+///
+///```dart
+///late final EmpireStringProperty name;
+///
+///name = createStringProperty('Bob');
+///```
+EmpireStringProperty createStringProperty(String value,
+    {String? propertyName}) {
+  return EmpireStringProperty(value, propertyName: propertyName);
+}
+
+///Short hand helper function for initializing an [EmpireNullableStringProperty].
+///
+///See [EmpireProperty] for [propertyName] usages.
+///
+///## Example
+///
+///```dart
+///late final EmpireNullableStringProperty name;
+///
+///name = createNullableStringProperty();
+///```
+EmpireNullableStringProperty createNullableStringProperty(
+    {String? value, String? propertyName}) {
+  return EmpireNullableStringProperty(value, propertyName: propertyName);
 }
